backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
openpyxl==3.1.5
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
reportlab>=4.2,<5
matplotlib>=3.9,<4
twilio>=9.3,<10
